fix(channels_sv2): migrate ShareAccounting from f64 to u64
Refactor share tracking fields to use u64 instead of f64 to avoid precision loss when accumulating shares.

Fixes #2092

+++ sv2/channels-sv2/src/client/extended.rs
@@ -217,7 +217,7 @@ impl<'a> ExtendedChannel<'a> {
     pub fn on_share_acknowledgement(
         &mut self,
         new_submits_accepted_count: u32,
-        new_shares_sum: f64,
+        new_shares_sum: u64,
     ) {
         self.share_accounting
             .on_share_acknowledgement(new_submits_accepted_count, new_shares_sum);
@@ -591,7 +591,7 @@ impl<'a> ExtendedChannel<'a> {
                 .track_validated_share(share.sequence_number, share_hash.to_raw_hash());
 
             // update the best diff
-            self.share_accounting.update_best_diff(share_hash_as_diff);
+            self.share_accounting.update_best_diff(share_hash_as_diff as u64);
 
             return Ok(ShareValidationResult::Valid(share_hash.to_raw_hash()));
         }
