+++ sv2/channels-sv2/src/client/standard.rs
@@ -173,7 +173,7 @@ impl<'a> StandardChannel<'a> {
     pub fn on_share_acknowledgement(
         &mut self,
         new_submits_accepted_count: u32,
-        new_shares_sum: f64,
+        new_shares_sum: u64,
     ) {
         self.share_accounting
             .on_share_acknowledgement(new_submits_accepted_count, new_shares_sum);
@@ -372,7 +372,7 @@ impl<'a> StandardChannel<'a> {
                 .track_validated_share(share.sequence_number, share_hash.to_raw_hash());
 
             // update the best diff
-            self.share_accounting.update_best_diff(share_hash_as_diff);
+            self.share_accounting.update_best_diff(share_hash_as_diff as u64);
 
             return Ok(ShareValidationResult::Valid(share_hash.to_raw_hash()));
         }
